fix: eliminate 2s connect_matrixone timeout in post-loop persist (no-pool fast path)
Root cause: persist_server_loop_core_events and all 5 Database*Writer
structs in turn/services.rs fell back to connect_matrixone() when
shared_pool was None, hitting a 2s acquire_timeout on every turn.

Fix:
- turn/services.rs: get_pool() is now sync and returns Err immediately
  when no pool is configured (fail-fast, not fallback-connect)
- run_lifecycle.rs: persist_server_loop_core_events early-returns when
  shared_pool is None (same guard as persist_runtime_promotion_events)
- server_loop_host.rs: fix 2 clippy unnecessary_cast warnings
- nextest.toml: set default profile slow-timeout to 15s
- common/mod.rs: fix doc_lazy_continuation clippy lint
- cargo fmt: format pre-existing unformatted code

Test: no_pool_writers_fail_fast_without_timeout asserts all 5 writers
complete in <100ms when pool is None (previously 2s each).

Impact: web_agent_e2e suite 30s → 6s, scenario_matrix 17s → 3s.

+++ rust/crates/runtime/src/server/server_loop_host.rs
@@ -1302,8 +1302,8 @@ impl ServerAgenticLoopHost {
 
         state.final_text_streamed = !full_text.is_empty();
         state.final_text = full_text;
-        state.total_prompt += u.input_tokens as u64;
-        state.total_completion += u.output_tokens as u64;
+        state.total_prompt += u.input_tokens;
+        state.total_completion += u.output_tokens;
         state.has_any_usage = true;
 
         Ok(HostTurnResult {
